Move command tests to their respective files
